requirements.txt
openai
whisper
torch
numpy
sounddevice
scipy
soundfile
webrtcvad
edge-tts
python-dotenv
